change data type

+++ src/Main.java
@@ -23,7 +23,7 @@ public class Main {
         return 0;
     }
 
-    public static double calcFeetAndInchesToCentimeters(int feet, int inches) {
+    public static double calcFeetAndInchesToCentimeters(double feet, double inches) {
         if(feet < 0 || inches < 0 || inches > 12) {
             return -1;
         }
@@ -32,12 +32,12 @@ public class Main {
         return centimeters;
     }
 
-    public static double calcFeetAndInchesToCentimeters(int inches) {
+    public static double calcFeetAndInchesToCentimeters(double inches) {
         if(inches < 0) {
             return -1;
         }
 
-        int feet = (inches - inches % 12)/ 12;
+        double feet = (inches - inches % 12)/ 12;
         inches = inches % 12;
         return calcFeetAndInchesToCentimeters(feet, inches);
     }
